requirements.txt
PyStemmer
psutil